% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{get_gt}
\alias{get_gt}
\title{Create a GT to display the sections for each connection}
\usage{
get_gt(selected_connection, title, subtitle)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
Create a GT to display the sections for each connection
}
