% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_connection.R
\name{mod_connection_ui}
\alias{mod_connection_ui}
\alias{mod_connection_server}
\title{mod_connection_ui and mod_connection_server}
\usage{
mod_connection_ui(id)

mod_connection_server(
  input,
  output,
  session,
  origin,
  departure,
  destination,
  arrival,
  sections
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
