% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_connections.R
\name{mod_connections_ui}
\alias{mod_connections_ui}
\alias{mod_connections_server}
\title{mod_connections_ui and mod_connections_server}
\usage{
mod_connections_ui(id)

mod_connections_server(input, output, session, origin, destination)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
