% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_connections_wrapper.R
\name{mod_connections_wrapper_ui}
\alias{mod_connections_wrapper_ui}
\alias{mod_connections_wrapper_server}
\title{mod_connections_wrapper_ui and mod_connections_wrapper_server}
\usage{
mod_connections_wrapper_ui(id)

mod_connections_wrapper_server(input, output, session, trip_details)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
