% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_trip_selector.R
\name{mod_trip_selector_ui}
\alias{mod_trip_selector_ui}
\alias{mod_trip_selector_server}
\title{mod_trip_selector_ui and mod_trip_selector_server}
\usage{
mod_trip_selector_ui(id)

mod_trip_selector_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
